+++ app/src/main/java/com/example/musicapp_gr2/MainActivity.java
@@ -1,4 +1,3 @@
-/*
 package com.example.musicapp_gr2;
 
 import android.os.Bundle;
@@ -25,53 +24,6 @@ public class MainActivity extends AppCompatActivity {
 
 
         });
-    }
-}*/
-
-package com.example.musicapp_gr2;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login); // Đảm bảo bạn đang dùng đúng file XML
-
-
-        // Tìm nút btnLoginPassword trong layout
-        Button btnLoginPassword = findViewById(R.id.btnLoginPassword);
-        Button tvSignUp = findViewById(R.id.tvSignUp);
-
-
-        // Kiểm tra nếu nút không null
-        if (btnLoginPassword != null) {
-            btnLoginPassword.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    // Chuyển sang SigninActivity
-                    Intent intent = new Intent(MainActivity.this, SigninActivity.class);
-                    startActivity(intent);
-                }
-            });
-        }
-       if (tvSignUp != null){
-           tvSignUp.setOnClickListener(new View.OnClickListener() {
-               @Override
-               public void onClick(View v) {
-                   Intent intent = new Intent(MainActivity.this, RegisterActivity.class);
-                   startActivity(intent);
-               }
-           });
-       }
     }
 }
